Refactor paint tool + implement eraser
